requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Async Support
aiofiles==23.2.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Image Processing
Pillow==10.2.0

# HTTP Client
httpx==0.26.0

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Moondream AI
moondream>=0.1.0

# Optional: Logging
loguru==0.7.2

